Rename repo to cl-github
I'm tired of the complaints from "clithub" while its a cute concat of cl
and github, quite a few people can't get past the "meaning" of it to see
that its a full github api.

Regardless the rename was a simple:

: sed -i s/\\bclithub\\b/cl-github/g *

in the repository, so maybe now everyone will be happy and I can quit
hearing stupid complaints about the name.

I'll release a new version to match the name change in a few days.